Moves some things into dfavm.h internal header
